requirements.txt
h5py==2.9.0
matplotlib==3.1.1
numpy==1.17.2
Pillow==6.2.0
Pillow==10.1.0
PyYAML==6.0.1
scikit_image==0.15.0
spikingjelly==0.0.0.0.14
tensorboardX==2.6.2.2
torch==1.3.1
torchvision==0.4.2
tqdm==4.36.1
